backend/requirements.txt
# Core LangChain and LangGraph (most stable, cache-friendly)
langchain
langgraph
langchain-google-genai
langgraph-sdk
langgraph-cli[inmem]
langgraph-api

# FastAPI and Web Framework (pinned for stability)
fastapi
uvicorn[standard]
websockets
python-multipart
starlette-compress

# LLM Providers (pinned versions)
google-genai
anthropic
openai
langchain-openai
langchain-groq
google-cloud-aiplatform

# Database and Storage (optimized)
sqlalchemy
alembic
asyncpg
psycopg2-binary
supabase
chromadb

# Redis and Caching
redis
aioredis

# Data Processing and Utils
python-dotenv
pydantic
aiohttp
python-jose[cryptography]
passlib[bcrypt]
PyJWT

# File Processing
agentic-doc
PyPDF2
docx2txt
python-docx
openpyxl
Pillow
fpdf

# Research and Analysis
pytrends
arxiv
beautifulsoup4
requests
feedparser

# Background Jobs
celery
celery[redis]

# Cloud Storage
boto3
azure-storage-blob

# Monitoring and Logging
prometheus-client
structlog
opentelemetry-api
opentelemetry-sdk

# Security
cryptography
bcrypt

# Development Tools
mypy
ruff
pytest
pytest-asyncio
httpx

# PyTorch (integrated wheel to avoid 2GB+ split CUDA packages)
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121

# ML and Embeddings
sentence-transformers